refactor: compoonent 분리 - 게임결과컴포넌트화 - 게임결과 사이드바 컴포넌트화 - data에 맞게끔 내용 필터링
