+++ src/app/(auth)/(with-header)/games/[gameId]/result/components/gameResultList/GameResultList.type.ts
@@ -1,0 +1,9 @@
+export interface GameResultListProps {
+  q: {
+    title: string;
+    question: string;
+    answer: string;
+    keywords: string[];
+  };
+  i: number;
+}
