+++ src/app/(auth)/(with-header)/games/[gameId]/result/components/gameResultList/index.ts
@@ -1,0 +1,1 @@
+export * from "./GameResultList";
