+++ src/app/(auth)/(with-header)/games/[gameId]/result/components/gameResultSidebar/GameResultSidebar.type.ts
@@ -1,0 +1,7 @@
+export interface GameResultData {
+  title: string;
+  question: string;
+  answer: string;
+  keywords: string[];
+  isCorrect: boolean;
+}
