+++ src/app/(auth)/(with-header)/games/[gameId]/result/components/gameResultSidebar/index.ts
@@ -1,0 +1,1 @@
+export * from "./GameResultSidebar";
